requirements-autotweet.txt
requests
tweepy
beautifulsoup4
pillow
google-generativeai
